clean up build warnings (debug x64 now 0 warnings)
- add /utf-8 to both projects -> fixes C4819 on korean comments
- make narrowing casts explicit (C4244): cos/sin/sqrt double->float,
  GetDesc().Width -> UINT, groupCount uint64->uint32
- fix UpdateSubresources misuse (returns byte count, 0=fail, not HRESULT)
- mark 2 headers as ClInclude (were ClCompile -> LNK4042/dup symbols)
- /ignore:4006 to silence third-party import-lib merge noise

+++ ZN/ZNFramework/Graphics/Platform/Direct3D12/DeferredLightingPass.cpp
@@ -341,7 +341,7 @@ void DeferredLightingPass::Render(GBufferManager* gbufferManager, ShadowMap* sha
     D3D12_CPU_DESCRIPTOR_HANDLE cpuHandle = lightingDescriptorHeap->GetCPUDescriptorHandleForHeapStart();
     D3D12_CONSTANT_BUFFER_VIEW_DESC cbvDesc = {};
     cbvDesc.BufferLocation = lightingConstantBuffer->GetGPUVirtualAddress();
-    cbvDesc.SizeInBytes = lightingConstantBuffer->GetDesc().Width;
+    cbvDesc.SizeInBytes = static_cast<UINT>(lightingConstantBuffer->GetDesc().Width);
     device->Device()->CreateConstantBufferView(&cbvDesc, cpuHandle);
 
     // Copy G-Buffer SRVs
